Logic: Find locations based on OpenAI response in proper way
Collect all city names encountered in the response. Collect them and
find the countries they reside in. Case that is not yet handled is when
no city names are written, but only countries. This comes in the next
commit :). The new logic is implemented in a different function called
`get_locations_serial()`. The old `get_locations()` is removed.

Add some tests along the way :).

Test plan:

Launching unittests with arguments:
```
$ python -m unittest platform/analyzer/analyzer.py
```